to double. some converter fixes

+++ assemblecpu.cpp
@@ -21,7 +21,7 @@ std::cout << "start assembling" << std::endl;
     for (size_t row{0}; row < num_nodes; ++row)
         matrix._rowptr[row + 1] = matrix._rowptr[row] + num_neighbours[row] + 1;
     matrix._colind = new size_t[matrix._rowptr[num_nodes]];
-    matrix._values = new float[matrix._rowptr[num_nodes]];
+    matrix._values = new double[matrix._rowptr[num_nodes]];
     for (size_t i{0}; i < matrix._rowptr[num_nodes]; ++i)
         matrix._values[i] = 0.0;
 
@@ -59,14 +59,14 @@ std::cout << "start assembling" << std::endl;
                 //        |  |
                 //       /    |
                 //      A------B
-                float B[2][2];
+                double B[2][2];
                 B[0][0] = nodes[nag[matrix._rowptr[row] - row + e    ]].x - nodes[row].x;
                 B[1][0] = nodes[nag[matrix._rowptr[row] - row + e    ]].y - nodes[row].y;
                 B[0][1] = nodes[nag[matrix._rowptr[row] - row + e + 1]].x - nodes[row].x;
                 B[1][1] = nodes[nag[matrix._rowptr[row] - row + e + 1]].y - nodes[row].y;
 //std::cout << B[0][0] << ", " << B[0][1] << ", " << B[1][0] << ", " << B[1][1] << ", " << std::endl;
-                float detB{std::abs(B[0][0]*B[1][1] - B[0][1]*B[1][0])};
-                float gradA[2], gradB[2], gradC[2]; // in fact these are the gradients multplied by det(B)
+                double detB{std::abs(B[0][0]*B[1][1] - B[0][1]*B[1][0])};
+                double gradA[2], gradB[2], gradC[2]; // in fact these are the gradients multplied by det(B)
                 gradA[0] = B[1][0] - B[1][1];
                 gradA[1] = B[0][1] - B[0][0];
                 gradB[0] = B[1][1];
@@ -83,13 +83,13 @@ std::cout << "start assembling" << std::endl;
 //                          << nag[matrix._rowptr[row] - row + e + 1] << ": " << ((gradA[0]*gradC[0] + gradA[1]*gradC[1]) / 2.0 / detB) << std::endl;
             }
             // e = num_neighbours[row] - 1; e + 1 = 0;
-                float B[2][2];
+                double B[2][2];
                 B[0][0] = nodes[nag[matrix._rowptr[row] - row + num_neighbours[row] - 1]].x - nodes[row].x;
                 B[1][0] = nodes[nag[matrix._rowptr[row] - row + num_neighbours[row] - 1]].y - nodes[row].y;
                 B[0][1] = nodes[nag[matrix._rowptr[row] - row /* e+1 = 0 */            ]].x - nodes[row].x;
                 B[1][1] = nodes[nag[matrix._rowptr[row] - row /* e+1 = 0 */            ]].y - nodes[row].y;
-                float detB(std::abs(B[0][0]*B[1][1] - B[0][1]*B[1][0]));
-                float gradA[2], gradB[2], gradC[2]; // in fact these are the gradients multplied by det(B)
+                double detB(std::abs(B[0][0]*B[1][1] - B[0][1]*B[1][0]));
+                double gradA[2], gradB[2], gradC[2]; // in fact these are the gradients multplied by det(B)
                 gradA[0] = B[1][0] - B[1][1];
                 gradA[1] = B[0][1] - B[0][0];
                 gradB[0] = B[1][1];
@@ -116,16 +116,16 @@ void assemble_cpu_elem(CsrMatrixCpu& matrix, std::vector<FullTriangle>& elements
 {
     for (const auto& elem : elements)
     {
-        float B[2][2];
+        double B[2][2];
         B[0][0] = elem.nodeB.x - elem.nodeA.x;
         B[1][0] = elem.nodeB.y - elem.nodeA.y;
         B[0][1] = elem.nodeC.x - elem.nodeA.x;
         B[1][1] = elem.nodeC.y - elem.nodeA.y;
-        //float d[2];
+        //double d[2];
         //d[0] = elem.nodeA.x;
         //d[1] = elem.nodeA.y;
-        float detB(std::abs(B[0][0]*B[1][1] - B[0][1]*B[1][0]));
-        float gradA[2], gradB[2], gradC[2]; // in fact these are the gradients multplied by det(B)
+        double detB(std::abs(B[0][0]*B[1][1] - B[0][1]*B[1][0]));
+        double gradA[2], gradB[2], gradC[2]; // in fact these are the gradients multplied by det(B)
         gradA[0] = B[1][0] - B[1][1];
         gradA[1] = B[0][1] - B[0][0];
         gradB[0] = B[1][1];
